+++ history.md
@@ -1,10 +1,10 @@
 ### TODOLIST
 - 实现tree-select
 - 实现多选的高度扩展设置项
-- FormView底部div指定模式下不合并！！！
 
 ### 2.1.29-beta
 - 依赖升级
+- FormView foot在inline模式下不能合并，Vue2不支持，因此合并FormFoot到FormView
 
 ### 2.1.28
 - 删除旧版本FormView